python: Rename `operational_webhook` to `operational_webhook_endpoint`
