+++ python/svix/api/__init__.py
@@ -60,7 +60,7 @@ from .message_attempt import (
     MessageAttemptListAttemptedDestinationsOptions,
     MessageAttemptResendOptions,
 )
-from .operational_webhook import (
+from .operational_webhook_endpoint import (
     OperationalWebhookEndpointAsync,
     OperationalWebhookEndpoint,
     OperationalWebhookEndpointListOptions,
